feat_14: review fixes
