+++ app/build.gradle.kts
@@ -67,6 +67,7 @@ dependencies {
     implementation(platform(libs.compose.bom))
     implementation(libs.compose.coil)
     implementation(libs.compose.ui)
+    implementation(libs.androidx.viewmodel)
 
     testImplementation(libs.junit4)
     testImplementation(libs.coroutines.test)
