--- app/src/main/java/ru/practicum/android/diploma/ui/common/ErrorChip.kt
@@ -1,59 +1,0 @@
-package ru.practicum.android.diploma.ui.common
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import ru.practicum.android.diploma.R
-
-@Composable
-fun ErrorChip(
-    visible: Boolean,
-    textId: Int,
-    iconId: Int,
-) {
-    if (visible)
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(535.dp)
-    ) {
-        Image(
-            modifier = Modifier
-                .align(Alignment.CenterHorizontally)
-                .padding(top = 156.dp),
-            painter = painterResource(iconId),
-            contentDescription = null
-        )
-        Text(
-            text = stringResource(textId),
-            style = TextStyle(
-                textAlign = TextAlign.Center,
-                color = MaterialTheme.colorScheme.onBackground,
-                fontFamily = FontFamily(
-                    fonts = listOf(Font(R.font.ys_display_medium))
-                ),
-                fontSize = 22.sp,
-                fontWeight = FontWeight(500)
-            ),
-            modifier = Modifier
-                .padding(top = 16.dp)
-                .align(Alignment.CenterHorizontally)
-        )
-    }
-}
