+++ app/src/main/res/drawable/ic_filter.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="12dp"
+    android:viewportWidth="18"
+    android:viewportHeight="12">
+  <path
+      android:pathData="M7,12H11V10H7V12ZM0,0V2H18V0H0ZM3,7H15V5H3V7Z"
+      android:fillColor="#1A1B22"/>
+</vector>
